improve low-snr performance

+++ src/counters.hpp
@@ -14,7 +14,7 @@
 class RxStatistics {
     uint32_t frames_received = 0;
     uint32_t frames_processed = 0;
-    std::complex<int8_t> dc_offset = {0, 0};
+    std::complex<float> dc_offset = {0, 0};
     float noise_power = 0;
     uint64_t last_reset_timestamp = 0;
 
@@ -22,7 +22,7 @@ class RxStatistics {
 
 public:
     void register_frame(bool processed);
-    void save_channel_characteristics(std::complex<int8_t> dc_offset, float noise_power);
+    void save_channel_characteristics(std::complex<float> dc_offset, float noise_power);
 
     void reset(uint64_t current_timestamp);
     bool reporting_due(uint64_t current_timestamp);
